requirements.txt
mss==10.1.0
pillow==12.2.0
tk==0.1.0
